Print subscribed regions as a string
Previously subscribed regions for reports were printed just passing the
internal lists `contagions_regional` and `vaccines_regional` to
`str.format`. This ends up like this:
```
["Abruzzo", "Basilicata", "Calabria"] --> 'Abruzzo', 'Basilicata',
'Calabria'
```

This commit converts the list into a string before passing it to
`helpers.bot.MyBot.send_message` as `fmt` argument, printing as
following:
```
["Abruzzo", "Basilicata", "Calabria"] --> Abruzzo, Basilicata,
Calabria
```

From user point of view, it change just removes quotation marks but
still this is not a list anymore in the code.