Infinite loop trying sending a message
Instead of try once and try again when a `ChatMigrated` error is
captured, use an infinite loop that stops only when the message is
actually sent.

Even if for a `ChatMigrated` error this is useless (since it is almost
impossible to get two consecutives chat migration), this will be very
useful to integrate other exceptions that could be raised more than
once.