Handle MarkdownV2 unescaped characters
MarkdownV2 have many common characters to be escaped (e.g. ".", "!",
"-", etc...). In `share/msg.*.md` files those characters can be escaped
by the author in the text but this is not possible when the message
requires a field to be inserted and formatted at runtime.
In this case author can not manage those texts and because of this a
programmtic check on text is desireable.

The automatic escape is done catching Telegram error and scraping the
unescaped character from exception's message.