Fix bug Italy considered as region
Commit a2d02be367b3038ccc098153e29bc0260f86fbf9 introduced as error
when Reporter removes "Italia" from the list of regions.
`list.pop` fails because it is misused, "Italia" remains in regions list
and it generates a fatal error while sending reports.

This commit fixes the bug replacing `list.pop` with `list.remove` and
replacing `except` with `except ValueError`; in this way if there is an
unexpected error it is not ignored as before.