/richiedi_report command
User can request a specific report on demand.

This new feature is implented using the existing code of
`/attiva_report` command. The output is not precise in some points,
reflecting the original purpose of the code (i.e. subscribing to
automatic updates, not answering to a request).

Some parts can be considered invalid but the operation goes as it has
to since original updates settings are restored after the end of
conversation.

In simple words, code works but output is not valid and should be
corrected maybe implementing a new part of code that satisfies only user
requests.